Avoid repeated messages for errors from asyn layer
This is done by remembering the last asynStatus in the AsynDriverInterface
object. In certain selected places, namely when pasynManager->queueRequest
or pasynCommon->connect is called, we report an error only if the new status
differs from the last one. This gives us just one message when the failure
occurs (or when the operation is tried for the first time) and another
message when the status returns back to asynSuccess.

Note this is currently limited to AsynDriverInterface. If a device is online
but responds in unexpected ways (including not at all) StreamDevice will
still report errors repeatedly whenever records are processed. Fixing this
is very well possible but beyond the scope of this patch.