Added compatibility for using vsnprintf on earlier versions of visual studio.
Under earlier versions of visual studio (VS2010 and below) vsnprintf will return a -1 if there is not enough room to print. Later versions (and linux) will return the number of characters that would have been printed if there was room, which was then used in StreamBuffer to expand the buffer. This change calls _vscprintf instead (for VS2010) to get the size required to expand the buffer.