Document stream-base.dbd usage when scalcout is not needed
Having StreamDevice built with scalcout support does not require that
the application also depend on calc if it does not use this record
type. But this is only true if we load `stream-base.dbd` instead of
`stream.dbd`. Therefore, this alternative database definition should be
documented in the setup page.